% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plot.R
\name{tile_plot}
\alias{tile_plot}
\title{Tile Plot over Multiple Columns}
\usage{
tile_plot(
  d = NULL,
  x = NULL,
  y = NULL,
  fill = NULL,
  facets = NULL,
  w = c(rep(0.94/max(d$col), max(d$col)), 0.06),
  y_label_free = FALSE,
  ...
)
}
\arguments{
\item{d}{Data set. Must contain a column \code{col} to indicate which \code{facet} is placed in which column of the plot}

\item{x}{Column in \code{d} that is plotted on x axis}

\item{y}{Column in \code{d} that is plotted on y axis}

\item{fill}{Column in \code{d} that used for the fill shading}

\item{facets}{Column in \code{d} that is used to create facets}

\item{w}{Spacing of columns}

\item{y_label_free}{Allow the space of the y labels in each column to vary. Will impact on the tile plotting area, creating unequal widths. Set to \code{FALSE} by default.}

\item{...}{Additional arguments passed to \code{facet_col()}}
}
\value{
A ggplot tile plot arranged into multiple columns
}
\description{
Tile Plot over Multiple Columns
}
\examples{
library(gapminder)
d <- gapminder \%>\%
  as_tibble() \%>\%
  mutate(col = as.numeric(continent),
         col = ifelse(test = continent == "Europe", yes = 2, no = col),
         col = ifelse(test = continent == "Oceania", yes = 3, no = col),
         # messing with label lengths to check code is working
         continent = fct_recode(continent, "North and South America" = "Americas"),
         country = as.character(country),
         country = ifelse(continent == "Africa", yes = abbreviate(country), country),
         country = fct_inorder(country),
         country = fct_rev(country))

tile_plot(d = d, x = "year", y = "country", fill = "lifeExp", facets = "continent")
tile_plot(d = d, x = "year", y = "country", fill = "lifeExp", facets = "continent",
          labeller = label_wrap_gen(width = 15))
}
